Move the news editor settings into a Post settings dialog.
The left pane now holds only the title, toolbar, and body field, with
the preview spanning the full form height. Summary, cover image, and
publish scheduling live in a new dialog (opened from a header button
next to Save, which also moved up from the bottom of the pane). The
dialog edits a local copy of the settings and applies it back on Done;
saving with a missing summary or invalid schedule opens the dialog,
since those fields are no longer on the page. Feedback from tec27.
